move frontend settings into the DB (for consistent settings handling).
Flattened settings object.

+++ webapp/backend/pkg/database/migrations/m20220716214900/setting.go
@@ -12,6 +12,6 @@ type Setting struct {
 	SettingKeyDescription string `json:"setting_key_description"`
 	SettingDataType       string `json:"setting_data_type"`
 
-	SettingValueNumeric int64  `json:"setting_value_numeric"`
+	SettingValueNumeric int    `json:"setting_value_numeric"`
 	SettingValueString  string `json:"setting_value_string"`
 }
